requirements.txt
click
pymysql
psycopg2-binary
python-dotenv
cryptography
mysql-connector-python

